Add files via upload
Investigators are interested in examining Type II Diabetes in 403 black participants residing in rural counties of Virginia. The data was combined from two published studies characterizing cardiovascular risk factors in Virginia for African Americans. The first study was a community-based study measuring the prevalence of coronary heart disease among Black population in Virginia. The second study measured the effectiveness of church-based smoking cessation interventions for rural Black populations. A smoking cessation program was implemented by church groups to target smoking behaviors. This program included one-on-one counseling with two counselors present, self-help materials, and community activities. Investigators followed-up with participants after 18 months to measure the effectiveness of the intervention. We now use this data to conduct studies to observe the relationship between the onset of Type II Diabetes and factors associated with the disease. Type II diabetes is known to be strongly associated with obesity and hypertension. The waist-to-hip ratio may also be a predictor of both diabetes and heart disease. Glycosylated hemoglobin (HbA1C) > 7.0 is usually indicative of diabetes. The goal of this paper is to describe the relationship between BMI and hemoglobin A1C and determine other factors that may impact this relationship. Researchers postulate that the higher the BMI the more likely an individual will have been diagnosed with diabetes.